# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Pick_User.Exclude_keyofUser.createdAt-or-updatedAt__": {
				"properties": {},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_User.createdAt-or-updatedAt_": {
				"$ref": "#/components/schemas/Pick_User.Exclude_keyofUser.createdAt-or-updatedAt__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"File": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"url": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"url"
				],
				"type": "object",
				"additionalProperties": false
			},
			"User": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"name": {
						"type": "string",
						"nullable": true
					},
					"surname": {
						"type": "string",
						"nullable": true
					},
					"photo": {
						"$ref": "#/components/schemas/File"
					},
					"wins": {
						"type": "number",
						"format": "double",
						"default": 0
					},
					"losses": {
						"type": "number",
						"format": "double",
						"default": 0
					},
					"draws": {
						"type": "number",
						"format": "double",
						"default": 0
					},
					"winRate": {
						"type": "number",
						"format": "double",
						"default": 0
					},
					"total": {
						"type": "number",
						"format": "double",
						"default": 0
					},
					"goals": {
						"type": "number",
						"format": "double",
						"default": 0
					},
					"assists": {
						"type": "number",
						"format": "double",
						"default": 0
					},
					"Elo": {
						"type": "number",
						"format": "double",
						"default": 0
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"username",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserInitResponse": {
				"properties": {
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"isNewUser": {
						"type": "boolean"
					}
				},
				"required": [
					"user",
					"isNewUser"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserUpdatePayload": {
				"properties": {
					"username": {
						"type": "string"
					},
					"name": {
						"type": "string",
						"nullable": true
					},
					"surname": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"username"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EEventStatus": {
				"enum": [
					"waiting",
					"started",
					"finished"
				],
				"type": "string"
			},
			"Location": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"preview": {
						"$ref": "#/components/schemas/File"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Event": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"dateTime": {
						"type": "string",
						"format": "date-time"
					},
					"price": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/EEventStatus"
					},
					"participantsLimit": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"location": {
						"$ref": "#/components/schemas/Location"
					},
					"participants": {
						"items": {
							"$ref": "#/components/schemas/Participant"
						},
						"type": "array"
					},
					"games": {
						"items": {
							"$ref": "#/components/schemas/Game"
						},
						"type": "array"
					},
					"teams": {
						"items": {
							"$ref": "#/components/schemas/Team"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"participants",
					"games",
					"teams",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Participant": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"waitList": {
						"type": "boolean"
					},
					"user": {
						"allOf": [
							{
								"$ref": "#/components/schemas/User"
							}
						],
						"nullable": true
					},
					"event": {
						"$ref": "#/components/schemas/Event"
					},
					"parentParticipant": {
						"$ref": "#/components/schemas/Participant"
					},
					"teamsParticipants": {
						"items": {
							"$ref": "#/components/schemas/TeamParticipant"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"waitList",
					"event",
					"teamsParticipants",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GameTeam": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"game": {
						"$ref": "#/components/schemas/Game"
					},
					"team": {
						"$ref": "#/components/schemas/Team"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"game",
					"team",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EGameEvent": {
				"enum": [
					"goal",
					"assist"
				],
				"type": "string"
			},
			"Game": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"event": {
						"$ref": "#/components/schemas/Event"
					},
					"gameTeams": {
						"items": {
							"$ref": "#/components/schemas/GameTeam"
						},
						"type": "array"
					},
					"gameStats": {
						"items": {
							"$ref": "#/components/schemas/GameStat"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"event",
					"gameTeams",
					"gameStats",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TeamParticipant": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"team": {
						"$ref": "#/components/schemas/Team"
					},
					"participant": {
						"$ref": "#/components/schemas/Participant"
					},
					"gameStats": {
						"items": {
							"$ref": "#/components/schemas/GameStat"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"team",
					"participant",
					"gameStats"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GameStat": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"event": {
						"$ref": "#/components/schemas/EGameEvent"
					},
					"game": {
						"$ref": "#/components/schemas/Game"
					},
					"teamParticipant": {
						"$ref": "#/components/schemas/TeamParticipant"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"event",
					"game",
					"teamParticipant",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Team": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"event": {
						"$ref": "#/components/schemas/Event"
					},
					"gameTeams": {
						"items": {
							"$ref": "#/components/schemas/GameTeam"
						},
						"type": "array"
					},
					"teamsParticipants": {
						"items": {
							"$ref": "#/components/schemas/TeamParticipant"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"event",
					"gameTeams",
					"teamsParticipants",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "sportstoryhubbl",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/api/user/all": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Omit_User.createdAt-or-updatedAt_"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/user/getById/{id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Omit_User.createdAt-or-updatedAt_"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/user/init": {
			"post": {
				"operationId": "UserInit",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IUserInitResponse"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"username": {
										"type": "string"
									}
								},
								"required": [
									"username"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/user/update": {
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserUpdatePayload"
							}
						}
					}
				}
			}
		},
		"/api/event/all": {
			"get": {
				"operationId": "GetAllEvents",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Event"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Events"
				],
				"security": [],
				"parameters": []
			}
		},
		"/api/event/getById/{id}": {
			"get": {
				"operationId": "GetEventById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Event"
								}
							}
						}
					}
				},
				"tags": [
					"Events"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/event/join-event": {
			"post": {
				"operationId": "JoinEvent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Events"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"username": {
										"type": "string"
									},
									"eventId": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"username",
									"eventId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/event/leave-event": {
			"delete": {
				"operationId": "LeaveEvent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Events"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"username": {
										"type": "string"
									},
									"eventId": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"username",
									"eventId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/location/all": {
			"get": {
				"operationId": "GetAllLocations",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Location"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Locations"
				],
				"security": [],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}